Added cross-compilation selection on github actions

+++ micro_ros_setup/config/freertos/olimex-stm32-e407/create.sh
@@ -2,11 +2,15 @@ pushd $FW_TARGETDIR >/dev/null
     # Install toolchain
     mkdir toolchain
     
-    echo "Downloading ARM compiler, this may take a while"
-    curl -fsSLO https://developer.arm.com/-/media/Files/downloads/gnu-rm/8-2019q3/RC1.1/gcc-arm-none-eabi-8-2019-q3-update-linux.tar.bz2 
-    tar --strip-components=1 -xvjf gcc-arm-none-eabi-8-2019-q3-update-linux.tar.bz2 -C toolchain  > /dev/null
-    rm gcc-arm-none-eabi-8-2019-q3-update-linux.tar.bz2
-    
+    if [[ ! -v GITHUB_ACTIONS ]]; then
+        # Install toolchain
+        mkdir toolchain
+
+        echo "Downloading ARM compiler, this may take a while"
+        curl -fsSLO https://developer.arm.com/-/media/Files/downloads/gnu-rm/8-2019q3/RC1.1/gcc-arm-none-eabi-8-2019-q3-update-linux.tar.bz2
+        tar --strip-components=1 -xvjf gcc-arm-none-eabi-8-2019-q3-update-linux.tar.bz2 -C toolchain  > /dev/null
+        rm gcc-arm-none-eabi-8-2019-q3-update-linux.tar.bz2
+    fi
     # Import repos
     vcs import --input $PREFIX/config/$RTOS/$PLATFORM/board.repos >/dev/null
